pyproject.toml
[project]
name = "trainer_agent"
version = "0.1.0"
description = "Multi-agent system for trainee task management using A2A and MCP"
requires-python = ">=3.9"
dependencies = [
    "python-a2a>=0.1.0",
    "mcp-python>=0.1.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pandas>=1.3.0",
    "beautifulsoup4>=4.9.3",
    "requests>=2.26.0",
    "python-dotenv>=0.19.0",
    "pydantic>=1.8.2",
    "aiohttp>=3.8.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."] 